+++ src/neptune/internal/disk_queue.py
@@ -33,6 +33,7 @@ from typing import (
 )
 
 from neptune.exceptions import MalformedOperation
+from neptune.internal.utils.files import remove_parent_folder_if_allowed
 from neptune.internal.utils.json_file_splitter import JsonFileSplitter
 from neptune.internal.utils.sync_offset_file import SyncOffsetFile
 
@@ -183,19 +184,7 @@ class DiskQueue(Generic[T]):
     def _remove_data(self):
         path = self._dir_path
         shutil.rmtree(path, ignore_errors=True)
-
-        parent = path.parent
-
-        try:
-            files = os.listdir(parent)
-        except FileNotFoundError:
-            files = []
-
-        if len(files) == 0:
-            try:
-                os.rmdir(parent)
-            except OSError:
-                _logger.info(f"Cannot remove directory: {parent}")
+        remove_parent_folder_if_allowed(path)
 
     def wait_for_empty(self, seconds: Optional[float] = None) -> bool:
         with self._empty_cond: