+++ tests/unit/neptune/new/attributes/atoms/test_file.py
@@ -36,6 +36,7 @@ from neptune.attributes.file_set import (
     FileSetVal,
 )
 from neptune.common.utils import IS_WINDOWS
+from neptune.envs import NEPTUNE_DISABLE_PARENT_DIR_DELETION
 from neptune.internal.operation import (
     UploadFile,
     UploadFileSet,
@@ -186,4 +187,17 @@ class TestFile(TestAttributeBase):
 
             run.stop()
 
-            assert not os.path.exists(data_path)
+            assert not os.path.exists(data_path)  # exec folder
+            assert not os.path.exists(data_path.parent)  # run folder
+
+    @patch.dict(os.environ, {NEPTUNE_DISABLE_PARENT_DIR_DELETION: "True"})
+    def test_clean_files_on_close_when_cleanup_disabled(self):
+        with self._exp() as run:
+            data_path = run._op_processor._operation_storage.data_path
+
+            assert os.path.exists(data_path)
+
+            run.stop()
+
+            assert not os.path.exists(data_path)  # exec folder
+            assert os.path.exists(data_path.parent)  # run folder