+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Fixes
 - Added more safe checking to last ack ([#1510](https://github.com/neptune-ai/neptune-client/pull/1510))
-
+- Retry request in case of bravado RecursiveCallException ([#1521](https://github.com/neptune-ai/neptune-client/pull/1512))
 
 ## neptune 1.8.2