+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [UNRELEASED] neptune 1.8.3
 
+### Changes
+- Safety (errors suppressing) execution mode ([#1503])(https://github.com/neptune-ai/neptune-client/pull/1503)
+- Allow to disable handling of remote signals ([#1508])(https://github.com/neptune-ai/neptune-client/pull/1508)
+- Allow to disable deletion of local parent folder ([#1511])(https://github.com/neptune-ai/neptune-client/pull/1511)
+
 ### Fixes
 - Added more safe checking to last ack ([#1510](https://github.com/neptune-ai/neptune-client/pull/1510))
 - Retry request in case of bravado RecursiveCallException ([#1521](https://github.com/neptune-ai/neptune-client/pull/1512))